+++ process_mining/src/conformance/mod.rs
@@ -4,5 +4,6 @@
 //! event data.
 pub mod case_centric;
 pub mod object_centric;
+#[allow(unused_imports)]
 pub use case_centric::*;
 pub use object_centric::*;